skip GTK detection if we're cross-compiling (#588)
It seems that stlink cannot use CMake's standard mechanism to find GTK3.
Cmake's GTK3 detection supports only Unix at the moment.

Therefore, stlink uses pkg-config to find GTK3. The downside of this is
that Cmake's cross-compiler settings are not taken into account. Even if
we're cross-compiling, CMake will find a local installation of GTK3 and
compilation of the GUI fails.

As a simple fix, we skip the GTK3 detection if we're cross-compiling.

(In the long run, we should use a more advanced mechanism to find GTK3.)