clear the PG bit before setting the PER bit
after the flash loader ran the PG bit is still set in the CR register.
in order to program the next page that page needs to be erased first and to get that working the PG bit has to be cleared so the flash controller knows what to do.
this fixes #579