Silence warning chipid files with STLINK_FLASH_TYPE_UNKNOWN
The chipid configuration parser currently emits a warning when it
encounters a file with the flash type set to
STLINK_FLASH_TYPE_UNKNOWN. In practice, this means that the warning
will always be printed since 'unknown_device.chip' uses this
value. Make STLINK_FLASH_TYPE_UNKNOWN a permitted value in the config
file parser.